templates/fastapi-app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==5.2.1
pycryptodome==3.20.0
